Update existing condition checks
