+++ geary-papermc-bridge/src/main/kotlin/com/mineinabyss/geary/papermc/bridge/conditions/Chance.kt
@@ -1,10 +1,21 @@
 package com.mineinabyss.geary.papermc.bridge.conditions
 
+import com.mineinabyss.geary.events.CheckingListener
+import com.mineinabyss.geary.systems.accessors.Pointers
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import org.bukkit.entity.Player
+import kotlin.random.Random
 
+@JvmInline
 @Serializable
-@SerialName("geary:condition.chance")
-data class Chance(
-    val percentage: Double
-)
+@SerialName("geary:chance")
+value class Chance(val percentage: Double)
+
+class ChanceChecker : CheckingListener() {
+    private val Pointers.chance by get<Chance>().on(source)
+
+    override fun Pointers.check(): Boolean {
+        return Random.nextDouble() < chance.percentage
+    }
+}
