+++ geary-papermc-bridge/src/main/kotlin/com/mineinabyss/geary/papermc/bridge/conditions/EntityConditions.kt
@@ -1,7 +1,11 @@
 package com.mineinabyss.geary.papermc.bridge.conditions
 
+import com.mineinabyss.geary.events.CheckingListener
+import com.mineinabyss.geary.papermc.bridge.helpers.nullOrEquals
+import com.mineinabyss.geary.systems.accessors.Pointers
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
+import org.bukkit.Material
 import org.bukkit.entity.LivingEntity
 import org.bukkit.entity.Player
 
@@ -33,5 +37,33 @@ class EntityConditions(
     val isInvurnerable: Boolean? = null,
     val isSilent: Boolean? = null,
     val isLeashed: Boolean? = null,
+)
 
-    )
+class EntityConditionsChecker : CheckingListener() {
+    private val Pointers.livingEntity by get<LivingEntity>().on(target)
+
+    private val Pointers.conditions by get<EntityConditions>().on(source)
+
+    override fun Pointers.check(): Boolean = livingEntity !is Player &&
+            conditions.isSleeping nullOrEquals livingEntity.isSleeping &&
+            conditions.isSwimming nullOrEquals livingEntity.isSwimming &&
+            conditions.isClimbing nullOrEquals livingEntity.isClimbing &&
+            conditions.isJumping nullOrEquals livingEntity.isJumping &&
+            conditions.isInLava nullOrEquals livingEntity.isInLava &&
+            conditions.isInWater nullOrEquals livingEntity.isInWater &&
+            conditions.isInBubbleColumn nullOrEquals livingEntity.isInBubbleColumn &&
+            conditions.isInRain nullOrEquals livingEntity.isInRain &&
+            conditions.isOnGround nullOrEquals livingEntity.isOnGround &&
+            conditions.isGliding nullOrEquals livingEntity.isGliding &&
+            conditions.isFrozen nullOrEquals livingEntity.isFrozen &&
+            conditions.isInPowderedSnow nullOrEquals livingEntity.isInPowderedSnow &&
+            conditions.isInCobweb nullOrEquals (livingEntity.location.block.type == org.bukkit.Material.COBWEB) &&
+            conditions.isInsideVehicle nullOrEquals livingEntity.isInsideVehicle &&
+            conditions.isRiptiding nullOrEquals livingEntity.isRiptiding &&
+            conditions.isInvisible nullOrEquals livingEntity.isInvisible &&
+            conditions.isGlowing nullOrEquals livingEntity.isGlowing &&
+            conditions.isInvurnerable nullOrEquals livingEntity.isInvulnerable &&
+            conditions.isSilent nullOrEquals livingEntity.isSilent &&
+            conditions.isLeashed nullOrEquals livingEntity.isLeashed
+
+}
