+++ geary-papermc-bridge/src/main/kotlin/com/mineinabyss/geary/papermc/bridge/events/entities/OnDamaged.kt
@@ -1,6 +1,8 @@
 package com.mineinabyss.geary.papermc.bridge.events.entities
 
 import com.mineinabyss.geary.papermc.bridge.events.EventHelpers
+import com.mineinabyss.geary.papermc.bridge.events.relations.Trigger
+import com.mineinabyss.geary.papermc.tracking.entities.toGeary
 import com.mineinabyss.geary.papermc.tracking.entities.toGearyOrNull
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
@@ -17,12 +19,18 @@ class EntityDamagedBridge : Listener {
     @EventHandler(ignoreCancelled = true)
     fun EntityDamageByEntityEvent.onDeath() {
         val gearyEntity = entity.toGearyOrNull() ?: return
-        EventHelpers.runSkill<OnDamaged>(gearyEntity)
+        val damager = damager.toGearyOrNull()
+        EventHelpers.runSkill<OnDamaged>(gearyEntity) {
+            damager?.let { addRelation<Trigger>(it) }
+        }
     }
 
     @EventHandler(ignoreCancelled = true)
     fun ProjectileHitEvent.onCollision() {
         val gearyEntity = hitEntity?.toGearyOrNull() ?: return
-        EventHelpers.runSkill<OnDamaged>(gearyEntity)
+        val damager = entity.toGearyOrNull()
+        EventHelpers.runSkill<OnDamaged>(gearyEntity) {
+            damager?.let { addRelation<Trigger>(it) }
+        }
     }
 }
